# Conflicts:
#	src/main/java/se/backend1/pensionat/controller/CustomerController.java
#	src/main/java/se/backend1/pensionat/dto/CustomerDto.java
#	src/main/java/se/backend1/pensionat/entity/Booking.java
#	src/main/java/se/backend1/pensionat/entity/Customer.java
#	src/main/java/se/backend1/pensionat/entity/Room.java
#	src/main/java/se/backend1/pensionat/mapper/CustomerMapper.java
#	src/main/java/se/backend1/pensionat/repository/CustomerRepository.java
#	src/main/java/se/backend1/pensionat/service/CustomerService.java
#	src/main/java/se/backend1/pensionat/service/impl/CustomerServiceImpl.java

+++ src/main/java/se/backend1/pensionat/repository/CustomerRepository.java
@@ -5,7 +5,8 @@ import se.backend1.pensionat.entity.Customer;
 
 import java.util.List;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> { // Tillgång till Crud-metoder mot SQL.DB
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+    // Tillgång till Crud-metoder mot SQL.DB
 
 //                //Mer specifika metoder utanför JpaRepository
 //    List<Customer> findByLastName(String lastName);
